Incomplete population of Form
My first attempt was to populate the dropdown menu from acsv file via
jquery.csv.js.  The population had issues, but I was able to print the
entire CSV file...
2nd attempt was to pretty much a hard-code creation of an HTML table and
then populate the drop down menu from that.  It was on the verge on
being functional, but not quite yet there.